Everything seems to finally be done. SFX added and play. Everything respects global mute.
Save data has a feature in it to allow it to be set with all old parameters so that when I add new settings if I do add them so that the old defaults can be utilized.

SFX play and can be muted. But if someone dies too quickly or holds the button down too fast then it won't play right but it works for now.

All that's left is adding copyright notices to the top of all of the source files.