Fixed some minor input bugs. Adjusted high score table. HighScores for the different modes weren't high enough for the top place. I played the game myself(I'm hardly an expert) and used my own scores after playing each.
I'm hardly an expert but the adjustments seem to have made them more inline with what should be a "high enough" score that it requires some skill to get but it's still not impossible. May shift them all up by 1k but that's up in the air.

Further in the text input of adding a score you'd mute the game if you pressed "m" that's been fixed by making sure that if you're in text input mode that no input is given to the states.